requirements.txt
# Core Dashboard & Data
streamlit
pandas
yfinance
requests
plotly
numpy
pytz

# Scheduling & Time
schedule
apscheduler
python-dotenv

# Web, Telegram, Email
flask
python-telegram-bot==13.15  # stable with legacy syntax
email-validator
beautifulsoup4
lxml

# Angel One API & WebSocket
websocket-client
pyotp

# AI Models & Strategies
scikit-learn
xgboost
prophet            # optional: time series forecast model (Facebook Prophet)

# Charts, Plots & Visuals
matplotlib
#seaborn
websocket